- Add new configuration for changing formatter of selection summary
