+++ packages/muze-utils/src/common-utils.js
@@ -1554,7 +1554,7 @@ const retrieveNearestGroupByReducers = (dataModel, ...measureFieldNames) => {
             getObjProp(derivations.find(derv => derv.op === DM_OPERATION_GROUP), 'criteria'), {});
 
         const measures = dataModel.getFieldspace().getMeasure();
-        measureFieldNames = measureFieldNames.length ? measureFieldNames: Object.keys(measures);
+        measureFieldNames = measureFieldNames.length ? measureFieldNames : Object.keys(measures);
         measureFieldNames.forEach((measureName) => {
             if (nearestReducers[measureName]) {
                 filteredReducers[measureName] = nearestReducers[measureName];
