+++ packages/muze-utils/src/enums/constants.js
@@ -1,1 +1,3 @@
 export const DM_OPERATION_GROUP = 'group';
+export const DM_OPERATION_COUNT = 'count';
+export const DM_OPERATION_SUM = 'sum';
