+++ packages/muze-utils/src/enums/index.js
@@ -1,3 +1,3 @@
 export { default as ReservedFields } from './reserved-fields';
 export { default as CommonProps } from './common-props';
-export { DM_OPERATION_GROUP } from './constants';
+export { DM_OPERATION_GROUP, DM_OPERATION_COUNT, DM_OPERATION_SUM } from './constants';
