+++ packages/muze-utils/src/index.js
@@ -16,6 +16,7 @@ export {
 } from './enums';
 
 export { STATE_NAMESPACES };
+export { DM_OPERATION_GROUP, DM_OPERATION_SUM, DM_OPERATION_COUNT } from './enums';
 
 export {
     scales,
